artifact3/enhanced/requirements.txt
dash
dash-leaflet
plotly
pymongo
pandas
numpy
matplotlib
python-dotenv